+++ ChangeLog.rst
@@ -1,5 +1,5 @@
-HEAD
-====
+0.24 - 2021-12-01
+=================
   - :class:`ihm.AsymUnit` now supports insertion codes in its
     ``auth_seq_id_map``. The target of this mapping can either be an
     author-provided residue number (as previously) or a 2-element tuple